requirements.txt
google-cloud-bigquery
sql-metadata
argparse
black